use single emitter
